third_party/WebKit/LayoutTests/http/tests/push_messaging/subscribe-failure-permission-denied-in-document.html

<!DOCTYPE html>
<html>
<head>
<title>subscribe() is rejected when permission has been denied</title>
<link rel="manifest" href="resources/push_manifest.json">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../serviceworker/resources/test-helpers.js"></script>
</head>
<body>
<script>
async_test(function(test) {
    var workerUrl = 'resources/empty_worker.js';
    var workerScope = 'resources/scope/' + location.pathname;
    var swRegistration;
    service_worker_unregister_and_register(test, workerUrl, workerScope)
        .then(function(serviceWorkerRegistration) {
            swRegistration = serviceWorkerRegistration;
            return wait_for_state(test, swRegistration.installing, 'activated');
        })
        .then(function() {
            // If running manually, deny permission when prompted.
            if (window.testRunner)
                testRunner.setPermission('push-messaging', 'denied', location.origin, location.origin);
            return swRegistration.pushManager.subscribe({ userVisibleOnly: true });
        })
        .then(function(pushSubscription) {
            assert_unreached('Subscription must not succeed after permission is denied.');
        }, function(e) {
            assert_equals(e.name, 'NotAllowedError');
            assert_regexp_match(e.message, /permission denied/);
            return service_worker_unregister_and_done(test, workerScope);
        })
        .catch(unreached_rejection(test));
}, 'subscribe() is rejected when permission has been denied');
</script>
</body>
</html>
