third_party/WebKit/LayoutTests/platform/win/http/tests/push_messaging/subscribe-failure-permission-denied-in-document-expected.txt

This is a testharness.js-based test.
PASS subscribe() is rejected when permission has been denied 
Harness: the test ran to completion.
